Sortert alt av models til Models prosjekt
