--- Gruppe10KVprototype/Models/Case.cs
@@ -1,22 +1,0 @@
-﻿/*using System.ComponentModel.DataAnnotations;
-
-namespace Gruppe10KVprototype.Models
-{
-    public class Case
-    {
-        [Key]
-        public int Id { get; set; }
-
-        public required string Username { get; set; }
-
-        public required string Description { get; set; }
-
-        public required string Status { get; set; }
-
-        public int GeoLocationId { get; set; }
-
-        // Add other relevant properties below.. This is currently just a template :)
-    }
-}
-*/
-
