--- Gruppe10KVprototype/Models/LandingPageModel.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace Gruppe10KVprototype.Models;
-
-public class LandingPageModel : Controller
-{
-    public IActionResult LandingPage()
-    {
-        return View();
-    }
-}
