+++ Gruppe10KVprototype/Views/UserPages/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model LoginModel
+﻿@model Models.InnmelderModels.InnmeldingLoginModel
 
 <body class="Index">
