+++ Gruppe10KVprototype/Views/UserPages/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model RegisterModel
+﻿@model Models.InnmelderModels.InnmelderRegisterModel
 @{
     ViewData["Title"] = "Register";
 }
