+++ Models/InnmelderModels/InnmelderLandingPageModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Models.InnmelderModels;
+
+public class InnmelderLandingPageModel : Controller
+{
+    public IActionResult LandingPage()
+    {
+        return View();
+    }
+}
