+++ Models/InnmelderModels/Innmelding.cs
@@ -1,0 +1,20 @@
+﻿/*using System.ComponentModel.DataAnnotations;
+
+namespace Models.InnmelderModels;
+public class Case
+{
+    [Key]
+    public int Id { get; set; }
+
+    public required string Username { get; set; }
+
+    public required string Description { get; set; }
+
+    public required string Status { get; set; }
+
+    public int GeoLocationId { get; set; }
+
+    // Add other relevant properties below.. This is currently just a template :)
+}
+*/
+
