+++ Models/InnmeldingModels/InnmeldingSaksbehandlerModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Models.InnmeldingModels;
 
-internal class InnmeldinSaksbehandlerModel
+internal class InnmeldingSaksbehandlerModel
 {
     public bool Avvist { get; set; }
 }
